move nimiq init logic to redux/actions
