--- shared/helpers/nimiq/index.js
@@ -1,3 +1,0 @@
-import nimiq from './nimiq'
-
-export { nimiq }
